test: extract Partition.test.ts (uses fake environment)
